Replaced get_create_sql with get_create_sql_dict
Changed get_table_names so it does not skip sqlite_sequence

Reason: If one ever wants to save all the tables in the db,
then you'll want to dump sqlite_sequence too. It keeps track
of the autoincrement status on all of your tables that have
this feature (i.e. it holds the value of the last id created for
each table with autoincrement).

Problem: Some developers won't know what this table does.
So, they might get confused if it shows up in the schema.